backend/profiling-service/requirements.txt
fastapi==0.110.*
uvicorn[standard]==0.29.*
pandas==2.2.*
openpyxl==3.1.*
PyJWT==2.8.*
